+++ python/kserve/kserve/handlers/liveness.py
@@ -1,0 +1,6 @@
+from kserve.handlers.base import BaseHandler
+
+
+class LivenessHandler(BaseHandler):  # pylint:disable=too-few-public-methods
+    def get(self):
+        self.write({"status": "alive"})
